[Refactor] file paths - authController.js , dashboardController.js
