+++ controllers/authController.js
@@ -1,6 +1,6 @@
 const path = require('path');
-const SignUp = require('../community/JS/Login/SignUpJson');
-const TimeStamp = require('../community/JS/Login/TimeStamp');
+const SignUp = require('../community/JS/Login/features/SignUpJson');
+const TimeStamp = require('../community/JS/Login/features/TimeStamp');
 const { v4 } = require('uuid');
