+++ controllers/dashboardController.js
@@ -1,7 +1,7 @@
 //대시보드 비즈니스 로직을 처리
 const path = require('path'); //디렉토리 경로를 다루는 기본 모듈
-const upload = require('../community/JS/DashBoard/multerConfig');
-const PostService = require('../community/JS/DashBoard/PostService');
+const upload = require('../community/JS/DashBoard/features/multerConfig');
+const PostService = require('../community/JS/DashBoard/features/PostService');
 
 //게시글 목록 조회
 exports.getDashboard = (req, res) => {
@@ -15,14 +15,14 @@ exports.getDashboardData = async(req,res) => {
         const posts = await PostService.getPosts(offset,limit);
 
         if(!posts){
-            res.status(400).json({'message': 'invalid_post'});
+            return res.status(400).json({'message': 'invalid_post'});
         }
 
-        res.status(200).json(posts);
+        return res.status(200).json(posts);
 
     }catch(error){
-        res.status(500).json({message : 'internal_server_error'});
         console.error('Error fetching dashboard data:', error);
+        return res.status(500).json({message : 'internal_server_error'});
     }
 }
 
@@ -38,7 +38,8 @@ exports.getPost = (req, res) => {
 
 //게시글 상세 조회 - 데이터 조회
 exports.getPostData = async (req,res) => {
-    let post_id = req.params.post_id;
+    //NOTE: params : url파라미터를 가져옴
+    const { post_id } = req.params;
 
     try{
         // NOTE: getPostById가 동기 함수여도 await는 정상 작동
@@ -47,10 +48,10 @@ exports.getPostData = async (req,res) => {
         const post = await PostService.getPostById(post_id); //id로 데이터 조회
 
         if(!post){
-            res.status(400).json({message : 'invalid_post_id'})
+            return res.status(400).json({message : 'invalid_post_id'})
         }
 
-        res.status(200).json(post);
+        return res.status(200).json(post);
 
     }catch(error){
         res.status(500).json({message : 'internal_server_error'});
@@ -87,10 +88,26 @@ exports.postAddPost = [upload.single('image'),(req,res) => {
         });
 
     }catch(error){
-        res.status(500).json({ message: 'server_error' });
         console.error('Error creating post:', error);
+        return res.status(500).json({ message: 'server_error' });
     }
-
-
-
 }];
+
+
+//게시글 삭제
+exports.deletePost = async (req,res)=>{
+    const { post_id } = req.params;
+
+    try{
+        const result = await PostService.deletePost(post_id);
+        if(!result) {
+            return res.status(400).json({message : 'invalid_post_id'});
+        }
+
+        return res.status(204).send();
+
+    }catch(error){
+        console.error('Error deleting post:', error);
+        return res.status(500).json({message : 'internal_server_error'});
+    }
+}
